Data File
You Can Use This Data To Make Predictions.
This Data is Taken From CPCB 
The Central Pollution Control Board of India is a statutory organization under the Ministry of Environment, Forest, and Climate Change. It was established in 1974 under the Water Act, 1974. The CPCB is also entrusted with the powers and functions under the Air Act, of 1981.